generate configs if all pods available
- this to avoid redundant sts-es reload
No reason to generate config if not all nodes are available.
In case of fresh deploy - this is unreal situation - nodes must be ready,
otherwise deploy cannot be finished at all.
In case of renew config - cluster is in degraded state, change of config in
this state is dangerous and may lead to broken cluster.

Change-Id: Id25a4b3de8b32044f0259febbaafcb4c23344722